% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mGJR.R
\name{mGJR}
\alias{mGJR}
\title{Provide the MGARCH-BEKK estimation procedure.}
\usage{
mGJR(eps1, eps2, order = c(1, 1, 1), params = NULL, fixed = NULL,
  method = "BFGS")
}
\arguments{
\item{eps1}{First time series.}

\item{eps2}{Second time series.}

\item{order}{mGJR(p, q, g) order}

\item{params}{Initial parameters for the \code{optim} function.}

\item{fixed}{Vector of parameters to be fixed.}

\item{method}{The method that will be used by the \code{optim} function.}
}
\value{
Estimation results packaged as \code{mGJR} class instance.
}
\description{
Provide the MGARCH-BEKK estimation procedure.
}

