% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEKK.R
\name{simulateBEKK}
\alias{simulateBEKK}
\title{Provides a procedure to simulate BEKK processes.}
\usage{
simulateBEKK(series.count, T, order = c(1, 1), params = NULL)
}
\arguments{
\item{series.count}{The number of series to be simulated.}

\item{T}{The length of series to be simulated.}

\item{order}{BEKK(p, q) order}

\item{params}{Initial parameters for simulation.}
}
\value{
Simulated series and auxiliary information packaged as a \code{simulateBEKK} class instance.
}
\description{
Provides a procedure to simulate BEKK processes.
}
\examples{
## Simulate series:
simulated = simulateBEKK(2, 1000, c(1,1))

}

